requirements.txt
base58==2.1.1
bidict==0.22.0
blockchain==1.4.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
enum-compat==0.0.3
Flask==2.2.2
Flask-SocketIO==5.3.1
future==0.18.2
idna==3.4
ipfshttpclient==0.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multiaddr==0.0.9
netaddr==0.8.0
pyAesCrypt==6.0.0
pycparser==2.21
python-engineio==4.3.4
python-socketio==5.7.2
requests==2.28.1
six==1.16.0
urllib3==1.26.12
varint==1.0.2
Werkzeug==2.2.2
